+++ src/components/adaptive-cards/AdaptiveCard/AdaptiveCard.jsx
@@ -3,16 +3,21 @@ import PropTypes from 'prop-types';
 import {Template} from 'adaptivecards-templating';
 import webexComponentClasses from '../../helpers';
 import AdaptiveCardContext from '../context/adaptive-card-context';
+import {mapValues} from '../../../util';
 
 import Component, {acPropTypes, registerComponent} from '../Component/Component';
 import '../ActionOpenURL/ActionOpenUrl';
 import '../ActionSet/ActionSet';
+import '../ActionShowCard/ActionShowCard';
+import '../ActionSubmit/ActionSubmit';
+import '../ActionToggleVisibility/ActionToggleVisibility';
 import '../Column/Column';
 import '../ColumnSet/ColumnSet';
 import '../Container/Container';
 import '../FactSet/FactSet';
 import '../Image/Image';
 import '../ImageSet/ImageSet';
+import '../InputChoiceSet/InputChoiceSet';
 import '../InputDate/InputDate';
 import '../InputNumber/InputNumber';
 import '../InputText/InputText';
@@ -55,6 +60,7 @@ AdaptiveCardInternal.defaultProps = {
 };
 
 AdaptiveCardInternal.acPropTypes = {
+  body: acPropTypes.children,
   type: acPropTypes.type,
   minHeight: acPropTypes.minHeight,
   rtl: acPropTypes.rtl,
@@ -71,6 +77,7 @@ registerComponent('AdaptiveCard', AdaptiveCardInternal, 'vertical');
  * @param {object} [props.context]  Provided data for binding to Adaptive Card
  * @param {string} [props.className]  Custom CSS class to apply
  * @param {object} [props.style]  Custom style to apply
+ * @param {Function} [props.onSubmit]  Action to perform on submit
  * @returns {object} JSX of the component
  */
 export default function AdaptiveCard({
@@ -78,12 +85,35 @@ export default function AdaptiveCard({
   context,
   className,
   style,
+  onSubmit,
 }) {
   const templateInstance = new Template(template);
   const data = templateInstance.expand({
     $root: context,
   });
+
   const [inputs, setInputs] = useState({});
+  const [elements, setElements] = useState({});
+
+  const setElement = useCallback((element) => {
+    setElements((prevElements) => (
+      {...prevElements, [element.id]: element}
+    ));
+  }, [setElements]);
+
+  const setIsVisible = (id, isVisible) => {
+    setElements((prevElements) => {
+      const targetElem = prevElements[id];
+
+      return {
+        ...prevElements,
+        [id]: {...targetElem, isVisible},
+      };
+    });
+  };
+
+  const getIsVisible = (id) => (elements[id]?.isVisible !== false);
+
   const setValue = (id, value) => {
     setInputs((prevInputs) => {
       const input = prevInputs[id];
@@ -94,21 +124,56 @@ export default function AdaptiveCard({
       };
     });
   };
+
   const setInput = useCallback((input) => {
     setInputs((prevInputs) => ({...prevInputs, [input.id]: input}));
   }, [setInputs]);
-  const getValue = (id, defval = '') => ((id in inputs && inputs[id].value) ? inputs[id].value : defval);
-  const getAllValues = () => Object.entries(inputs).reduce((allValues, [id, input]) => (
-    {...allValues, [id]: input.value}
-  ), {});
+
+  const getValue = (id, defval = '') => ((id in inputs && inputs[id].value !== undefined) ? inputs[id].value : defval);
+
+  const getAllValues = () => mapValues(inputs, (input) => (input.value));
+
+  const getError = (id) => inputs[id]?.error;
+
+  const validate = () => {
+    const newInputs = mapValues(inputs, (input) => {
+      let error;
+
+      if (input.isRequired && !input.value && input.value !== 0) {
+        error = input.errorMessage;
+      } else if (input.value < input.min) {
+        error = `Minimum value is ${input.min}`;
+      } else if (input.value > input.max) {
+        error = `Maximum value is ${input.max}`;
+      } else if (input.regex && !String(input.value).match(input.regex)) {
+        error = input.errorMessage;
+      } else if (String(input.value).length > input.maxLength) {
+        error = `Maximum length is ${input.maxLength}`;
+      }
+
+      return {...input, error};
+    });
+
+    setInputs(newInputs);
+
+    return Object.values(newInputs).some((input) => input.error);
+  };
+
+  const submit = (values) => onSubmit(values);
 
   return (
     <AdaptiveCardContext.Provider
       value={{
         setValue,
         getValue,
-        setInput,
         getAllValues,
+        setInput,
+        getError,
+        validate,
+        submit,
+        setElement,
+        setIsVisible,
+        getIsVisible,
       }}
     >
       <Component data={data} className={className} style={style} />
@@ -121,10 +186,12 @@ AdaptiveCard.propTypes = {
   context: PropTypes.shape(),
   className: PropTypes.string,
   style: PropTypes.shape(),
+  onSubmit: PropTypes.func,
 };
 
 AdaptiveCard.defaultProps = {
   className: undefined,
   context: undefined,
   style: undefined,
+  onSubmit: undefined,
 };