+++ src/components/adaptive-cards/Container/Container.jsx
@@ -40,6 +40,7 @@ Container.defaultProps = {
 Container.acPropTypes = {
   backgroundImage: acPropTypes.backgroundImage,
   bleed: acPropTypes.bleed,
+  fallback: acPropTypes.fallback,
   height: acPropTypes.height,
   id: acPropTypes.id,
   isVisible: acPropTypes.isVisible,
@@ -54,6 +55,7 @@ Container.acPropTypes = {
 };
 
 Container.acDefaultProps = {
+  isVisible: true,
   verticalContentAlignment: 'top',
 };